Refactor existing tests into CRUD
