+++ src/test/java/seedu/address/logic/commands/delete/ClearContactCommandTest.java
@@ -1,11 +1,10 @@
-package seedu.address.logic.commands.contact;
+package seedu.address.logic.commands.delete;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.TypicalContacts.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.delete.ClearContactCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
