+++ src/test/java/seedu/address/logic/commands/update/EditContactCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.contact;
+package seedu.address.logic.commands.update;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -18,7 +18,6 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.update.EditContactCommand;
 import seedu.address.logic.commands.update.EditContactCommand.EditContactDescriptor;
 import seedu.address.logic.commands.delete.ClearContactCommand;
 import seedu.address.model.Model;
