+++ src/test/java/seedu/address/logic/parser/todo/TodoParserTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.create.AddTodoCommand;
 import seedu.address.logic.commands.delete.DeleteTodoCommand;
-import seedu.address.logic.commands.read.DisplayTodoInformationCommand;
+import seedu.address.logic.commands.read.InfoTodoCommand;
 import seedu.address.logic.commands.read.ListTodoCommand;
 import seedu.address.logic.commands.update.MarkTodoAsDoneCommand;
 import seedu.address.logic.commands.update.MarkTodoAsNotDoneCommand;
@@ -40,10 +40,10 @@ public class TodoParserTest {
 
     @Test
     public void parseCommand_displayInfo() throws Exception {
-        DisplayTodoInformationCommand command = (DisplayTodoInformationCommand) parser.parseCommand(
-                TODO_COMMAND_WORD + " " + DisplayTodoInformationCommand.COMMAND_WORD + " "
+        InfoTodoCommand command = (InfoTodoCommand) parser.parseCommand(
+                TODO_COMMAND_WORD + " " + InfoTodoCommand.COMMAND_WORD + " "
                         + INDEX_FIRST.getOneBased());
-        assertEquals(new DisplayTodoInformationCommand(INDEX_FIRST), command);
+        assertEquals(new InfoTodoCommand(INDEX_FIRST), command);
     }
 
     @Test
